requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
email-validator==1.1.3
SQLAlchemy==1.4.23
WTForms==2.3.3
Werkzeug==2.0.1
python-dotenv==0.19.0
Flask-Migrate==3.1.0
python-dateutil==2.8.2
beautifulsoup4==4.10.0
requests==2.26.0 